+++ cookbooks/vm/recipes/base.rb
@@ -1,9 +1,8 @@
 
 include_recipe 'apt'
-include_recipe 'chef-sugar'
 
 # commonly needed packages / tools
-%w(vim libxml2-dev libxslt1-dev zlib1g-dev liblzma-dev build-essential xvfb indicator-multiload).each do |pkg|
+%w(vim libcurl3 gconf2 libxml2-dev libxslt1-dev zlib1g-dev liblzma-dev build-essential xvfb indicator-multiload).each do |pkg|
   package pkg
 end
 
@@ -13,6 +12,7 @@ directory "#{vm_user_home}/Desktop" do
   group vm_user
   mode '0755'
 end
+
 cookbook_file "#{vm_user_home}/Desktop/README.md" do
   source "desktop_readme.md"
   owner vm_user