+++ cookbooks/vm/spec/integration/recipes/ansible_spec.rb
@@ -10,16 +10,16 @@ describe 'vm::ansible' do
   it 'installs ansible 2.3.0.0' do
     expect(ansible_version).to contain '2.3.0.0'
   end
-  it 'installs ansible-lint 3.4.12' do
-    expect(ansible_lint_version).to contain '3.4.12'
+  it 'installs ansible-lint 3.4.23' do
+    expect(ansible_lint_version).to contain '3.4.23'
   end
-  it 'installs testinfra 1.6.3' do
-    expect(pytest_version).to contain 'testinfra-1.6.3'
+  it 'installs testinfra 1.16.0' do
+    expect(pytest_version).to contain 'testinfra-1.16.0'
   end
   it 'installs pytest-spec 1.1.0' do
     expect(pytest_version).to contain 'pytest-spec-1.1.0'
   end
-  it 'installs molecule 2.0.0.0rc6' do
-    expect(molecule_version).to contain '2.0.0.0rc6'
+  it 'installs molecule 2.19.0' do
+    expect(molecule_version).to contain '2.19.0'
   end
 end