+++ cookbooks/vm/spec/integration/recipes/atom_spec.rb
@@ -6,22 +6,22 @@ describe 'vm::atom' do
   let(:atom_config) { file("#{vm_user_home}/.atom/config.cson") }
   let(:installed_plugins) { vm_user_command('apm list -i -b').stdout }
 
-  it 'installs atom 1.17.1' do
-    expect(atom_version).to contain '1.17.1'
+  it 'installs atom 1.32.2' do
+    expect(atom_version).to contain '1.32.2'
   end
 
   describe 'plugins' do
-    it 'installs "atom-beautify" plugin v0.29.24' do
-      expect(installed_plugins).to contain 'atom-beautify@0.29.24'
+    it 'installs "atom-beautify" plugin v0.33.4' do
+      expect(installed_plugins).to contain 'atom-beautify@0.33.4'
     end
-    it 'installs "minimap" plugin v4.28.2' do
-      expect(installed_plugins).to contain 'minimap@4.28.2'
+    it 'installs "minimap" plugin v4.29.9' do
+      expect(installed_plugins).to contain 'minimap@4.29.9'
     end
-    it 'installs "language-chef" plugin v0.9.0' do
-      expect(installed_plugins).to contain 'language-chef@0.9.0'
+    it 'installs "language-chef" plugin v3.5.1' do
+      expect(installed_plugins).to contain 'language-chef@3.5.1'
     end
-    it 'installs "language-ansible" plugin v0.2.1' do
-      expect(installed_plugins).to contain 'language-ansible@0.2.1'
+    it 'installs "language-ansible" plugin v0.2.2' do
+      expect(installed_plugins).to contain 'language-ansible@0.2.2'
     end
   end