+++ cookbooks/vm/spec/integration/recipes/vagrant_spec.rb
@@ -5,8 +5,8 @@ describe 'vm::vagrant' do
   let(:vagrant_version) { vm_user_command('vagrant -v').stdout.strip }
   let(:installed_plugins) { vm_user_command('vagrant plugin list').stdout }
 
-  it 'installs vagrant 1.9.5' do
-    expect(vagrant_version).to match 'Vagrant 1.9.5'
+  it 'installs vagrant 2.2.0' do
+    expect(vagrant_version).to match 'Vagrant 2.2.0'
   end
 
   if in_vmware?
@@ -21,22 +21,22 @@ describe 'vm::vagrant' do
 
   describe 'plugins' do
     it 'installs "vagrant-cachier" plugin v1.2.1' do
-      expect(installed_plugins).to include 'vagrant-cachier (1.2.1)'
+      expect(installed_plugins).to include 'vagrant-cachier (1.2.1, global)'
     end
-    it 'installs "vagrant-berkshelf" plugin v5.1.1' do
-      expect(installed_plugins).to include 'vagrant-berkshelf (5.1.1)'
+    it 'installs "vagrant-berkshelf" plugin v5.1.2' do
+      expect(installed_plugins).to include 'vagrant-berkshelf (5.1.2, global)'
     end
     it 'installs "vagrant-omnibus" plugin v1.5.0' do
-      expect(installed_plugins).to include 'vagrant-omnibus (1.5.0)'
+      expect(installed_plugins).to include 'vagrant-omnibus (1.5.0, global)'
     end
     it 'installs "vagrant-toplevel-cookbooks" plugin v0.2.4' do
-      expect(installed_plugins).to include 'vagrant-toplevel-cookbooks (0.2.4)'
+      expect(installed_plugins).to include 'vagrant-toplevel-cookbooks (0.2.4, global)'
     end
     it 'installs "vagrant-managed-servers" plugin v0.8.0' do
-      expect(installed_plugins).to include 'vagrant-managed-servers (0.8.0)'
+      expect(installed_plugins).to include 'vagrant-managed-servers (0.8.0, global)'
     end
-    it 'installs "vagrant-lxc" plugin v1.2.3' do
-      expect(installed_plugins).to include 'vagrant-lxc (1.2.3)'
+    it 'installs "vagrant-lxc" plugin v1.4.3' do
+      expect(installed_plugins).to include 'vagrant-lxc (1.4.3, global)'
     end
   end
 end