+++ cookbooks/vm/spec/integration/scripts/update_vm_spec.rb
@@ -2,8 +2,8 @@ require 'spec_helper'
 
 describe 'update-vm.sh' do
 
-  it 'installs chefdk 1.4.3' do
-    expect(file('/opt/chefdk/version-manifest.txt')).to contain 'chefdk 1.4.3'
+  it 'installs chefdk 3.4.38' do
+    expect(file('/opt/chefdk/version-manifest.txt')).to contain 'chefdk 3.4.38'
   end
 
   it 'symlinks the update-vm script to /usr/local/bin/' do