+++ CHANGELOG.md
@@ -4,10 +4,18 @@
 
 Updated tools:
 
+ * updated basebox to Ubuntu 18.04 (see [PR #68](https://github.com/tknerr/linus-kitchen/pull/68)
+    * optimized performance of VM in Virtualbox
+    * update to ChefDK v3.4.23 and cookbooks
+    * update to Vagrant v2.2.0 and plugins
+    * update to Atom v1.32.2 and plugins
+    * update to Docker v18.06.1
+    * update to Packer v1.3.2
+    * update to VirtualBox v5.2.22
  * updated the ansible toolchain (see [PR #66](https://github.com/tknerr/linus-kitchen/pull/66))
     * added python-vagrant wrapper so we can use molecule's vagrant driver
-    * update to molecule v2.0.0.rc6
-    * update to testinfra v1.6.3
+    * update to molecule v2.19.0
+    * update to testinfra v1.16.0
 
 ## 0.3 (May 30, 2017)